pyproject.toml
[project]
name = "dativo-ingest"
version = "1.1.0"
description = "Dativo ingestion runner - config-driven data ingestion engine"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "dagster>=1.5.0",
    "python-json-logger>=2.0.0",
    "jsonschema>=4.0.0",
    "requests>=2.28.0",
    "pandas>=2.0.0",
    "pyarrow>=14.0.0",
    "pyiceberg>=0.5.0",
    "pynessie>=0.60.0",
    "boto3>=1.28.0",
    "psycopg2-binary>=2.9.0",
    "mysql-connector-python>=8.0.0",
    "stripe>=7.0.0",
    "hvac>=2.1.0",
    "google-cloud-secret-manager>=2.20.0",
    "docker>=6.0.0",
    "vcrpy>=5.0.0",
    "google-api-python-client>=2.0.0",
    "google-auth>=2.0.0",
    "metadata-ingestion>=0.13.0",
    "databricks-sdk>=0.18.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-mock>=3.0.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["dativo_ingest"]

[tool.setuptools.package-dir]
"" = "src"

[project.scripts]
dativo = "dativo_ingest.cli:main"
dativo-ingest = "dativo_ingest.cli:main"

[tool.black]
line-length = 88
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88

